Update `css_variables_exporter.py` to use cli arguments and move scripts to `scripts` folder
