--- misc/requirements.txt
@@ -1,2 +1,0 @@
-tinycss2
-webcolors
