+++ packages/config-eslint/moss-lint-plugin/css_variables.json
@@ -48,9 +48,47 @@
   "--moss-red-9",
   "--moss-red-10",
   "--moss-red-11",
+  "--moss-yellow-1",
+  "--moss-yellow-2",
+  "--moss-yellow-3",
+  "--moss-yellow-4",
+  "--moss-yellow-5",
+  "--moss-yellow-6",
+  "--moss-yellow-7",
+  "--moss-yellow-8",
+  "--moss-yellow-9",
+  "--moss-yellow-10",
+  "--moss-orange-1",
+  "--moss-orange-2",
+  "--moss-orange-3",
+  "--moss-orange-4",
+  "--moss-orange-5",
+  "--moss-orange-6",
+  "--moss-orange-7",
+  "--moss-orange-8",
+  "--moss-orange-9",
+  "--moss-purple-1",
+  "--moss-purple-2",
+  "--moss-purple-3",
+  "--moss-purple-4",
+  "--moss-purple-5",
+  "--moss-purple-6",
+  "--moss-purple-7",
+  "--moss-purple-8",
+  "--moss-purple-9",
+  "--moss-teal-1",
+  "--moss-teal-2",
+  "--moss-teal-3",
+  "--moss-teal-4",
+  "--moss-teal-5",
+  "--moss-teal-6",
+  "--moss-teal-7",
+  "--moss-teal-8",
+  "--moss-teal-9",
   "--moss-primary",
   "--moss-error",
   "--moss-error-background",
+  "--moss-success",
   "--moss-success-background",
   "--moss-border-color",
   "--moss-primary-background",
@@ -58,17 +96,30 @@
   "--moss-primary-text",
   "--moss-secondary-background",
   "--moss-secondary-background-hover",
+  "--moss-secondary-background-active",
   "--moss-secondary-text",
   "--moss-info-background",
   "--moss-info-background-hover",
   "--moss-info-text",
   "--moss-info-icon",
   "--moss-info-border",
+  "--moss-divider-color",
+  "--moss-not-selected-item-color",
+  "--moss-shortcut-text",
+  "--focus-border",
   "--separator-border",
   "--moss-icon-primary-background",
   "--moss-icon-primary-text",
   "--moss-icon-primary-background-hover",
   "--moss-icon-primary-background-active",
+  "--moss-icon-secondary-background",
+  "--moss-icon-secondary-text",
+  "--moss-icon-secondary-background-hover",
+  "--moss-icon-secondary-background-active",
+  "--moss-headBar-icon-primary-text",
+  "--moss-headBar-primary-background",
+  "--moss-headBar-primary-background-hover",
+  "--moss-headBar-border-color",
   "--moss-statusBar-icon-primary-text",
   "--moss-statusBar-icon-secondary-text",
   "--moss-statusBar-icon-background-hover",
@@ -79,6 +130,7 @@
   "--moss-windowControlsLinux-text",
   "--moss-windowControlsLinux-hoverBackground",
   "--moss-windowControlsLinux-activeBackground",
+  "--moss-sash-active-highlight",
   "--moss-paneview-active-outline-color",
   "--moss-paneview-header-border-color",
   "--moss-drag-over-background-color",
@@ -92,6 +144,7 @@
   "--moss-inactivegroup-visiblepanel-tab-color",
   "--moss-activegroup-visiblepanel-tab-border-color",
   "--moss-activegroup-hiddenpanel-tab-border-color",
+  "--moss-inactivegroup-visiblepanel-tab-border-color",
   "--moss-activegroup-hiddenpanel-tab-background-color",
   "--moss-inactivegroup-hiddenpanel-tab-background-color",
   "--moss-activegroup-hiddenpanel-tab-color",
@@ -101,53 +154,40 @@
   "--moss-floating-box-shadow",
   "--moss-overlay-z-index",
   "--moss-button-primary-solid-background",
+  "--moss-button-primary-solid-background-hover",
+  "--moss-button-primary-solid-background-active",
   "--moss-button-primary-solid-border",
+  "--moss-button-primary-solid-border-hover",
+  "--moss-button-primary-solid-border-active",
   "--moss-button-primary-solid-text",
-  "--moss-button-primary-solid-boxShadow",
   "--moss-button-primary-outlined-background",
+  "--moss-button-primary-outlined-background-hover",
+  "--moss-button-primary-outlined-background-active",
   "--moss-button-primary-outlined-border",
+  "--moss-button-primary-outlined-border-hover",
+  "--moss-button-primary-outlined-border-active",
   "--moss-button-primary-outlined-text",
-  "--moss-button-primary-outlined-boxShadow",
-  "--moss-button-primary-soft-background",
-  "--moss-button-primary-soft-border",
-  "--moss-button-primary-soft-text",
-  "--moss-button-primary-soft-boxShadow",
-  "--moss-button-primary-ghost-background",
-  "--moss-button-primary-ghost-border",
-  "--moss-button-primary-ghost-text",
-  "--moss-button-primary-ghost-boxShadow",
-  "--moss-button-danger-solid-background",
-  "--moss-button-danger-solid-border",
-  "--moss-button-danger-solid-text",
-  "--moss-button-danger-solid-boxShadow",
-  "--moss-button-danger-outlined-background",
-  "--moss-button-danger-outlined-border",
-  "--moss-button-danger-outlined-text",
-  "--moss-button-danger-outlined-boxShadow",
-  "--moss-button-danger-soft-background",
-  "--moss-button-danger-soft-border",
-  "--moss-button-danger-soft-text",
-  "--moss-button-danger-soft-boxShadow",
-  "--moss-button-danger-ghost-background",
-  "--moss-button-danger-ghost-border",
-  "--moss-button-danger-ghost-text",
-  "--moss-button-danger-ghost-boxShadow",
   "--moss-button-neutral-solid-background",
+  "--moss-button-neutral-solid-background-hover",
+  "--moss-button-neutral-solid-background-active",
   "--moss-button-neutral-solid-border",
+  "--moss-button-neutral-solid-border-hover",
+  "--moss-button-neutral-solid-border-active",
   "--moss-button-neutral-solid-text",
-  "--moss-button-neutral-solid-boxShadow",
   "--moss-button-neutral-outlined-background",
+  "--moss-button-neutral-outlined-background-hover",
+  "--moss-button-neutral-outlined-background-active",
   "--moss-button-neutral-outlined-border",
+  "--moss-button-neutral-outlined-border-hover",
+  "--moss-button-neutral-outlined-border-active",
   "--moss-button-neutral-outlined-text",
-  "--moss-button-neutral-outlined-boxShadow",
-  "--moss-button-neutral-soft-background",
-  "--moss-button-neutral-soft-border",
-  "--moss-button-neutral-soft-text",
-  "--moss-button-neutral-soft-boxShadow",
-  "--moss-button-neutral-ghost-background",
-  "--moss-button-neutral-ghost-border",
-  "--moss-button-neutral-ghost-text",
-  "--moss-button-neutral-ghost-boxShadow",
+  "--moss-button-background-disabled",
+  "--moss-button-background-disabled-hover",
+  "--moss-button-background-disabled-active",
+  "--moss-button-border-disabled",
+  "--moss-button-border-disabled-hover",
+  "--moss-button-border-disabled-active",
+  "--moss-button-text-disabled",
   "--moss-controls-plain-bg",
   "--moss-controls-plain-border",
   "--moss-controls-plain-text",
@@ -159,6 +199,8 @@
   "--moss-select-border-outlined",
   "--moss-select-text-outlined",
   "--moss-select-hover-bg",
+  "--moss-select-item-bg-outlined-hover",
+  "--moss-select-item-bg-outlined-selected",
   "--moss-input-bg-plain",
   "--moss-input-bg-outlined",
   "--moss-stepCard-text",
