+++ scripts/svg_component_generator.py
